extracting package io into separate file
