+++ lib/init.js
@@ -1,8 +1,6 @@
-var fs = require('fs')
-var Path = require('path')
 var addressRegex = require('bitcoin-regex')({ exact: true })
-var sortObject = require('sorted-object')
 var extend = require('xtend')
+var packageLoader = require('./package-loader')
 
 module.exports = function (address, cb) {
   if (!cb) {
@@ -17,7 +15,7 @@ module.exports = function (address, cb) {
   var cwd = process.cwd()
 
   // read existing package.json
-  read(cwd, function (err, json) {
+  packageLoader.read(cwd, function (err, json) {
     if (err) {
       if (err.code === 'ENOENT') {
         return cb(new Error('No package.json in current directory.'))
@@ -31,27 +29,6 @@ module.exports = function (address, cb) {
       address: address
     })
 
-    write(cwd, json, cb)
+    packageLoader.write(cwd, json, cb)
   })
-}
-
-function read (dir, cb) {
-  fs.readFile(
-    Path.join(dir, 'package.json'),
-    'utf8',
-    function (err, contents) {
-      if (err) { return cb(err) }
-      cb(null, JSON.parse(contents))
-    }
-  )
-}
-
-function write (dir, json, cb) {
-  json.sustain = sortObject(json.sustain)
-
-  fs.writeFile(
-    Path.join(dir, 'package.json'),
-    JSON.stringify(json, null, 2),
-    cb
-  )
 }
